Add toMap config utility functions
